@@ -18,8 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
-  flutter: "2.17.6"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -35,16 +35,17 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  google_fonts: ^2.1.0
-  equatable: ^2.0.3
-  flutter_bloc: ^8.0.0
-  intl: ^0.17.0
+  google_fonts: ^6.1.0
+  equatable: ^2.0.5
+  flutter_bloc: ^8.1.3
+  intl: ^0.19.0
   step_progress_indicator: ^1.0.2
-  firebase_auth: ^3.3.6
-  firebase_core: ^1.12.0
-  image_picker: ^0.8.5
-  firebase_storage: ^10.2.15
-  cloud_firestore: ^3.1.14
+  firebase_auth: ^4.15.3
+  firebase_core: ^2.24.2
+  image_picker: ^1.0.5
+  firebase_storage: ^11.5.6
+  cloud_firestore: ^4.13.6
+  js: ^0.6.3
 
 dev_dependencies:
   flutter_test:
