@@ -1,15 +1,19 @@
 part of 'auth_bloc.dart';
 
 @immutable
-abstract class AuthEvent extends Equatable{
+abstract class AuthEvent extends Equatable {
   const AuthEvent();
   @override
   List<Object?> get props => [];
 }
 
-class AuthUserChanged extends AuthEvent{
+class AuthUserChanged extends AuthEvent {
   final auth.User? user;
   const AuthUserChanged({required this.user});
   @override
   List<Object?> get props => [user];
+}
+
+class AuthLogoutRequested extends AuthEvent {
+  const AuthLogoutRequested();
 }
