@@ -1,21 +1,25 @@
-part of 'profile_bloc.dart';
+import 'package:equatable/equatable.dart';
+
+import '../../model/user_model.dart';
 
 abstract class ProfileState extends Equatable {
   const ProfileState();
-
   @override
   List<Object?> get props => [];
 }
 
-class ProfileLoading extends ProfileState {
-  @override
-  List<Object?> get props => [];
-}
+class ProfileLoading extends ProfileState {}
 
-class ProfileLoaded extends ProfileState{
-final User user;
-
+class ProfileLoaded extends ProfileState {
+  final User user;
   const ProfileLoaded({required this.user});
   @override
   List<Object?> get props => [user];
+}
+
+class ProfileError extends ProfileState {
+  final String message;
+  const ProfileError({required this.message});
+  @override
+  List<Object?> get props => [message];
 }
