@@ -7,6 +7,7 @@ class SignupState extends Equatable {
   final String password;
   final SignupStatus status;
   final auth.User? user;
+  final String? errorMessage;
 
   bool get isFormValid => email.isNotEmpty && password.isNotEmpty;
 
@@ -14,30 +15,36 @@ class SignupState extends Equatable {
     required this.email,
     required this.password,
     required this.status,
-     this.user,
+    this.user,
+    this.errorMessage,
   });
 
   factory SignupState.initial() {
     return const SignupState(
-        email: '',
-        password: '',
-        status: SignupStatus.initial,
-        user: null);
+      email: '',
+      password: '',
+      status: SignupStatus.initial,
+      user: null,
+      errorMessage: null,
+    );
   }
 
   @override
-  List<Object?> get props => [email, password, status, user];
+  List<Object?> get props => [email, password, status, user, errorMessage];
 
   SignupState copyWith({
     String? email,
     String? password,
     SignupStatus? status,
     auth.User? user,
+    String? errorMessage,
   }) {
     return SignupState(
-        email: email ?? this.email,
-        password: password ?? this.password,
-        status: status ?? this.status,
-        user: user ?? this.user);
+      email: email ?? this.email,
+      password: password ?? this.password,
+      status: status ?? this.status,
+      user: user ?? this.user,
+      errorMessage: errorMessage,
+    );
   }
 }
